# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: AI Homework Grader API
  version: "0.2.0"
  description: |
    API for AI-powered homework grading. Upload exam photos, receive annotated
    grading results, and browse error-question collections. MVP covers English
    and Math.

    ## Authentication (v0.2.0)
    Parent identity is established via SMS verification-code login
    (`POST /api/auth/send-code` + `POST /api/auth/login`), which issues a JWT.
    All business endpoints require `Authorization: Bearer <token>`; the JWT
    `sub` claim is the parent id (userId). Phone appears ONLY in the auth
    endpoints — never in business URLs, query params, or logs.

    Image endpoints use signed URLs (HMAC token + expiry in the query string)
    because `<img>`/`<image>` tags cannot send an Authorization header.

    The WeChat mini-program uses the SAME SMS login flow as the Web client
    (`POST /api/auth/send-code` + `POST /api/auth/login`); there is no
    mini-program-specific login endpoint. The issued JWT is cached client-side
    and reused until it expires.
servers:
  - url: http://localhost:8000
    description: Local development (FastAPI backend)

# Global security: all business endpoints require a Bearer JWT unless overridden.
security:
  - bearerAuth: []
...
paths:
...
  /api/error-collections:
    get:
      operationId: listErrorQuestions
      tags: [Error Collections]
      summary: List error questions with filters
      parameters:
        - name: child_id
          in: query
          required: false
          schema:
            type: integer
        - name: subject
          in: query
          required: false
          schema:
            type: string
            enum: [english, math]
        - name: question_type
          in: query
          required: false
          schema:
            type: string
            enum: [choice, fill_blank, reading, composition, calculation, word_problem]
        - name: from_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Filtered error questions
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ErrorQuestion"
                  total:
                    type: integer
  /api/error-collections/generate:
    post:
      operationId: generateErrorSheet
      tags: [Error Collections]
      summary: Generate a practice sheet from error questions
      description: |
        Generates a practice sheet from the parent's error questions. Two
        formats:

        - `image` (default, backward compatible): composes cropped question
          images into a single sheet image (existing behavior).
        - `text`: assembles transcribed question text (English text / Math
          LaTeX) into a text sheet and produces a downloadable .docx. Math
          LaTeX is rendered to PNG and embedded in the .docx. Sheets do NOT
          include answer keys.

        The frontend sheet-generation page defaults to "text" and sends
        `format=text` explicitly; the API default remains `image` for backward
        compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [child_id, subject]
              properties:
                child_id:
                  type: integer
                subject:
                  type: string
                  enum: [english, math]
                question_types:
                  type: array
                  items:
                    type: string
                    enum: [choice, fill_blank, reading, composition, calculation, word_problem]
                from_date:
                  type: string
                  format: date
                to_date:
                  type: string
                  format: date
                count:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 50
                format:
                  type: string
                  enum: [text, image]
                  default: image
      responses:
        "200":
          description: Generated practice sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedSheet"

# ═══════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════
components:
  schemas:
...
    ErrorQuestion:
      type: object
      required:
        - id
        - submission_id
        - child_id
        - child_name
        - subject
        - question_number
        - question_type
        - question_image_path
        - error_count
        - last_error_at
        - created_at
      properties:
        id:
          type: integer
        submission_id:
          type: integer
        child_id:
          type: integer
        child_name:
          type: string
        subject:
          type: string
          enum: [english, math]
        question_number:
          type: string
        question_type:
          type: string
          enum: [choice, fill_blank, reading, composition, calculation, word_problem]
        question_image_path:
          type: string
          description: Signed URL to the cropped error question region image
        question_text:
          type: string
          nullable: true
          description: Transcribed full question text (English). Redundant from GradedQuestion.
        question_latex:
          type: string
          nullable: true
          description: Transcribed question in LaTeX (Math). Redundant from GradedQuestion.
        solution_note:
          type: string
          nullable: true
        error_category:
          type: string
          nullable: true
          enum: [grammar, vocabulary, spelling, logic, calculation, careless, comprehension]
        error_count:
          type: integer
          description: Number of times this question was answered incorrectly (within same submission)
        error_timestamps:
          type: array
          items:
            type: string
            format: date-time
          description: Timestamps of each error occurrence
        is_manually_fixed:
          type: boolean
        last_error_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # ── Sheet Generation ──────────────────────────────
    GeneratedSheet:
      type: object
      required: [format, question_count]
      properties:
        format:
          type: string
          enum: [text, image]
        question_count:
          type: integer
        image_url:
          type: string
          format: uri
          description: Present when format=image. Signed URL to the generated sheet image.
        questions:
          type: array
          items:
            $ref: "#/components/schemas/SheetQuestion"
          description: Present when format=text. Randomly selected error questions.
        docx_url:
          type: string
          format: uri
          description: Present when format=text. Signed URL to download the .docx sheet.
    SheetQuestion:
      type: object
      required: [question_number, question_type, subject]
      properties:
        question_number:
          type: string
        question_type:
          type: string
          enum: [choice, fill_blank, reading, composition, calculation, word_problem]
        subject:
          type: string
          enum: [english, math]
        question_text:
          type: string
          nullable: true
          description: Transcribed text (English); rendered directly.
        question_latex:
          type: string
          nullable: true
          description: Transcribed LaTeX (Math); rendered client-side (KaTeX) or as PNG in .docx.
        question_image_path:
          type: string
          nullable: true
          description: Signed URL to a fallback cropped image for questions without usable text.
        source_submission_id:
          type: integer
          description: Origin submission (for traceability).

    # ── Generic ───────────────────────────────────────
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by POST /api/auth/login. The `sub` claim is the parent id
        (userId). Applies to all business endpoints.
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Future scheme — retained for migration reference only. NOT implemented
        in v2.

  # ── Responses ───────────────────────────────────────
...